refactor: move the catalog entry schema into ipc/types
McpCatalogEntrySchema lived in the main-process catalog client, so the
IPC contract's value import pulled that module (electron-log, fetch and
cache state) into the renderer bundle. Move the schema and its type to a
dependency-free file under ipc/types and import it from both places.
